Updated the last two paragraphs with Beta content
From November 2019 to April 2020, the team publicly deployed a **Beta** version of the National Cybercrime and Fraud Reporting System to gain clarity around how this service will be used by victims and to build capacity with an RCMP delivery team that will be taking over the iterative design and development of the service.

Heading into Summer 2020, the CDS and RCMP team will move into a Transition Phase, where we will look to shift the service to 100% RCMP ownserhip. Along with this, we will take the learnings from the pilot and work to plan what the end-to-end full version of the service will be, and how we get there.